backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch==2.2.0
torchvision==0.17.0
transformers==4.36.0
Pillow==10.1.0
numpy<2.0,>=1.24.0  # Must be < 2.0 for PyTorch compatibility
requests==2.31.0
python-dotenv==1.0.0
setuptools>=65.0
opencv-python>=4.8.0
ultralytics>=8.0.0
ollama>=0.1.0